+++ src/pygerber/examples/__init__.py
@@ -16,6 +16,19 @@ class ExamplesEnum(Enum):
     simple_2layer_F_Paste = "simple_2layer-F_Paste.gbr"  # noqa: N815
     simple_2layer_F_Silkscreen = "simple_2layer-F_Silkscreen.gbr"  # noqa: N815
 
+    carte_test_B_Cu = "carte_test-B_Cu.gbr"  # noqa: N815
+    carte_test_B_Mask = "carte_test-B_Mask.gbr"  # noqa: N815
+    carte_test_B_Paste = "carte_test-B_Paste.gbr"  # noqa: N815
+    carte_test_B_Silkscreen = "carte_test-B_Silkscreen.gbr"  # noqa: N815
+
+    carte_test_Edges = "carte_test-Edge_Cuts.gbr"  # noqa: N815
+
+    carte_test_F_Cu = "carte_test-F_Cu.gbr"  # noqa: N815
+    carte_test_F_Mask = "carte_test-F_Mask.gbr"  # noqa: N815
+    carte_test_F_Silkscreen = "carte_test-F_Silkscreen.gbr"  # noqa: N815
+
+    carte_test_gbrjob = "carte_test-job.gbrjob"
+
 
 DIRECTORY = Path(__file__).parent