+++ src/pygerber/gerber/api/_project.py
@@ -16,9 +16,30 @@ class Project:
         self,
         *,
         top: CompositeView,
-        internal_layers: Iterable[CompositeView],
+        inner: Iterable[CompositeView],
         bottom: CompositeView,
     ) -> None:
-        self.top = top
-        self.internal_layers = tuple(internal_layers)
-        self.bottom = bottom
+        self._top = top
+        self._inner = tuple(inner)
+        self._bottom = bottom
+
+    @property
+    def top(self) -> CompositeView:
+        """Get top view."""
+        return self._top
+
+    @property
+    def inner(self) -> tuple[CompositeView, ...]:
+        """Get inner views."""
+        return self._inner
+
+    @property
+    def bottom(self) -> CompositeView:
+        """Get bottom view."""
+        return self._bottom
+
+    def __str__(self) -> str:
+        return (
+            f"{self.__class__.__qualname__}(top={self._top}, "
+            f"inner={self._inner}, bottom={self._bottom})"
+        )