+++ test/examples/gerberx3/api/test_examples.py
@@ -50,3 +50,27 @@ def test_quickstart_examples(example_path: Path) -> None:
 def test_single_file_examples(example_path: Path) -> None:
     with cd_to_tempdir():
         exec(example_path.read_text(encoding="utf-8"))  # noqa: S102
+
+
+@pytest.mark.parametrize(
+    "example_path",
+    [
+        *THIS_DIRECTORY.glob("*.no_output.py"),
+    ],
+    ids=lambda path: path.name,
+)
+def test_no_output_examples(example_path: Path) -> None:
+    with cd_to_tempdir():
+        exec(example_path.read_text(encoding="utf-8"))  # noqa: S102
+
+
+def test_multi_layer_gerber_job() -> None:
+    from pygerber.gerber.api._gerber_job_file import GerberJobFile
+    from test.assets.gerberx3.A64_OLinuXino_rev_G import A64_OlinuXino_Rev_G
+
+    gerber_job = GerberJobFile.from_file(A64_OlinuXino_Rev_G.gbrjob.path)
+    project = gerber_job.to_project()
+
+    assert len(project.top) != 0
+    assert len(project.inner) != 0
+    assert len(project.bottom) != 0